moved tum file parsing into io.cpp for everyone to use
